refactor: apply supabase.config instead of database.config
- Delete database.config.js
- Rename passport.js to passport.config.js
