--- backend/src/config/database.js
@@ -1,21 +1,0 @@
-const { createClient } = require('@supabase/supabase-js');
-const logger = require('../utils/logger');
-
-const supabase = createClient(
-  process.env.SUPABASE_URL,
-  process.env.SUPABASE_ANON_KEY
-);
-
-const testConnection = async () => {
-  try {
-    const { data, error } = await supabase.from('users').select('count').limit(1);
-
-    if (error) throw error;
-    logger.info('Connect to Supabase successfully');
-  } catch (error) {
-    logger.error('Connect to Supabase failed', error.message);
-  }
-};
-testConnection();
-
-module.exports = supabase;
