+++ backend/src/models/user.model.js
@@ -1,4 +1,4 @@
-const supabase = require('../config/database');
+const { supabase } = require('../config/supabase.config');
 
 class UserModel {
   async findById(id) {
